requirements.txt
---
## requirements.txt
```txt
qiskit
qiskit-aer
pillow
numpy